requirements.txt
blinker==1.9.0
cachelib==0.13.0
click==8.2.1
colorama==0.4.6
et_xmlfile==2.0.0
Flask==3.1.1
Flask-Excel==0.0.7
Flask-Session==0.8.0
itsdangerous==2.2.0
Jinja2==3.1.6
lml==0.2.0
MarkupSafe==3.0.2
msgspec==0.19.0
mysql-connector-python==9.3.0
openpyxl==3.1.5
pyexcel==0.7.3
pyexcel-io==0.6.7
pyexcel-webio==0.1.4
pyexcel-xlsx==0.6.0
texttable==1.7.0
Werkzeug==3.1.3
